add storage.rs under cluster/raft to separate raft domain logic from storage
