+++ src/clusters/raft/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod messages;
 pub(crate) mod state;
 pub(crate) mod transport;
 pub(crate) mod multi_raft;
+pub(crate) mod storage;
