+++ src/storage/mod.rs
@@ -72,27 +72,6 @@ pub(crate) enum DbOp {
     },
 }
 
-/// we place the following functions under WriteOperation so that Raft details don't leak into Storage layer
-impl DbOp {
-    pub(crate) fn put_log_entry(cf: &str, entry: &LogEntry) -> Self {
-        let key = ShardCfKey::LogEntry(entry.index).encode();
-        let value = bincode::encode_to_vec(entry, BINCODE_CONFIG).expect("encode LogEntry failed");
-        DbOp::Put { cf: cf.to_owned(), key, value }
-    }
-
-    pub(crate) fn put_hard_state(cf: &str, term: u64, voted_for: Option<NodeId>) -> Self {
-        let key = ShardCfKey::HardState.encode();
-        let value = bincode::encode_to_vec(&(term, voted_for), BINCODE_CONFIG)
-            .expect("encode HardState failed");
-        DbOp::Put { cf: cf.to_owned(), key, value }
-    }
-
-    pub(crate) fn delete_from(cf: &str, from_index: u64) -> Self {
-        let index = ShardCfKey::LogEntry(from_index).encode();
-        DbOp::DeleteFrom { cf: cf.to_owned(), index }
-    }
-}
-
 /// Opaque handle to the node's RocksDB instance.
 /// Callers interact only through this API — `rocksdb` does not leak outside this module.
 pub(crate) struct Db {
